+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Added
+
+- Added more Sandpile demos and more content to the Sandpile tutorial in the docs
+
+## Changed
+
+- Changed interpretation of the `cellular_automaton` argument in `evolve` and `evolve2d` such that a history of states can be provided
 
 ## [2.0.0] - 2021-11-10